requirements.txt
fastapi==0.128.8
pydantic==2.12.5
python-dotenv==1.0.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
uvicorn==0.39.0
